refactor: schema is now called dto
