+++ gallagher/cc/alarms/schedule.py
@@ -3,13 +3,14 @@
 """
 
 from ..utils import APIBase
-from ...schema.schedule import ScheduleSummaryResponse
+from ...dto.schedule import ScheduleSummaryResponse
+
 
 class Schedule(APIBase):
     """ Schedules
     """
 
     class Config:
-        
+
         endpoint = "schedules"
         list_response_class = ScheduleSummaryResponse
