+++ gallagher/dto/operator.py
