+++ gallagher/dto/utils.py
